+++ .formatter.exs
@@ -12,6 +12,8 @@
     answer: 3,
     answer: 4,
     edit: 5,
+    keyboard: 2,
+    keyboard: 3,
     row: 1,
     button: 1,
     button: 2,
@@ -27,6 +29,7 @@
   exports: [
     locals_without_parens: [
       keyboard: 2,
+      keyboard: 3,
       button: 1,
       button: 2,
       reply_button: 1,